feat: Introduce Accessibility and Clipboard Managers for enhanced functionality
- Added AccessibilityManager to handle accessibility permissions and status monitoring.
- Introduced ClipboardManager for managing clipboard operations and auto-paste functionality.
- Refactored AudioRecorder to delegate accessibility and clipboard operations to their respective managers.
- Updated ConfigurationCard to conditionally display manual permissions check based on accessibility status.
- Improved overall code organization and maintainability by separating concerns into dedicated manager classes.
